1. added median for mysql, as udf 2. upgraded worthy scoring

+++ datasys/crawler_helper.py
@@ -18,12 +18,12 @@ def persist_db_history_and_latest(table_name, num_cols, value_list, is_many=True
     affected_rows = 99999
     if need_history:
         sql = 'replace into %s values(%s)' %(table_name,values_str)
-        affected_rows = dbhelper.executeSqlWriteMany(sql,value_list)
+        affected_rows = dbhelper.executeSqlWriteMany(sql,value_list,is_dirty=True)
         t2 = time.time()
         logging.debug('persist_db_history_and_latest, history using time: %s' %(t2-t1))
     t22 = time.time()
     sql2 = 'replace into %s values(%s)' %(tbl_latest,values_str)
-    affected_rows2 = dbhelper.executeSqlWriteMany(sql2,value_list)
+    affected_rows2 = dbhelper.executeSqlWriteMany(sql2,value_list,is_dirty=True)
     t3 = time.time()
     logging.debug('persist_db_history_and_latest, latest using time: %s' %(t3-t22))
     status = -1
@@ -36,7 +36,7 @@ def persist_db_history_and_latest(table_name, num_cols, value_list, is_many=True
     }
     return ret
 
-def persist_db_history_and_lastest_empty_first(table_name, num_cols, value_list, is_many=True, need_history=False):
+def persist_db_history_and_lastest_empty_first(table_name, num_cols, value_list, is_many=True, need_history=False, sql_create_table=None):
 
     ret = {
         'status':-1,
@@ -51,7 +51,11 @@ def persist_db_history_and_lastest_empty_first(table_name, num_cols, value_list,
         ps_list.append('%s')
     values_str = ','.join(ps_list)
 
-    sql_empty = 'delete from %s' %tbl_latest
+    sql_empty = 'drop table if exists %s' %tbl_latest
+    if sql_create_table is None:
+        sql_empty = 'delete from %s' %tbl_latest
+    print sql_empty
+
     sql = 'replace into %s values(%s)' %(table_name,values_str)
     sql2 = 'replace into %s values(%s)' %(tbl_latest,values_str)
 
@@ -59,12 +63,20 @@ def persist_db_history_and_lastest_empty_first(table_name, num_cols, value_list,
     affected_rows = 0
     try:
         cursor1 = conn.cursor()
+        cursor1.execute('set @@session.tx_isolation="serializable"')
+        t1 = time.time()
         afr1 = cursor1.execute(sql_empty)
+        t2 = time.time()
+        afr15 = -1
+        if sql_create_table is not None:
+            afr15 = cursor1.execute(sql_create_table)
         afr2 = cursor1.executemany(sql2, value_list)
+        t3 = time.time()
         afr3 = 0
         if need_history:
             afr3 = cursor1.executemany(sql,value_list)
         conn.commit()
+        t4 = time.time()
         affected_rows = cursor1.rowcount
         ret = {
             'status':0,
@@ -73,6 +85,9 @@ def persist_db_history_and_lastest_empty_first(table_name, num_cols, value_list,
             'affected_rows':afr3
         }
         print 'Rows deleted: %s\nRows added latest: %s\nRows added: %s' %(afr1,afr2,afr3)
+        print 'delete data using seconds: %0.1f' %(t2-t1)
+        print 'inserting data using seconds: %0.1f' %(t3-t2)
+        print 'commit using seconds: %0.1f' %(t4-t3)
     except Exception as e:
         conn.rollback()
         print e
